SecureMessenger/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
cryptography
PyNaCl
python-jose[cryptography]
python-dotenv
aiosqlite
passlib[bcrypt]
requests
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
PyJWT==2.9.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
